Precreate fixes in obj_spell_handler
Renamed obj_spell to obj_spell_handler with the intention of having one global spell handler object that handles spells for all casters.
Moved expressions for obj_spell_handler inside of the precreate event to the create event to get around gamemaker bug.
Syntatic corrections made for referencing keys in a ds_map inside of ds_lists for the activation and cooldown timers.